Fix Solutions
For the add functionality we don't want to search the user's cart, we want to search through the list of swag. If we do find the product with the given id, we want to add it to the the user's cart. The `.find` is unnecessary as well since we already have the index.